app/requirements.txt
fastapi
uvicorn
websockets
sqlmodel
python-jose[cryptography]
passlib[bcrypt]
python-multipart
openai
python-dotenv
jieba
python-docx
pdfplumber
PyMuPDF
Pillow
numpy
paddleocr
paddlepaddle
paddlex[ocr]==3.4.3
PyJWT==2.8.0
passlib==1.7.4
bcrypt==4.1.2
jieba==0.42.1
openpyxl==3.1.2
xlrd==2.0.1
redis
langgraph
langchain-openai
langchain-core
chromadb
sentence-transformers
aiosqlite
langgraph-checkpoint-sqlite
pytest-asyncio
PyYAML
